# docs/openapi/SCIMServer_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SCIMServer API",
    "version": "0.37.3",
    "description": "Complete SCIM 2.0 Server - 82 endpoints across 19 controllers.\n\nMulti-tenant endpoint architecture, OAuth 2.0, per-endpoint credentials, bulk operations, /Me endpoint, custom resource types, database browser, activity feed, per-endpoint logs, structured logging.\n\nAll SCIM paths are scoped under /endpoints/{endpointId}. Admin paths are under /admin.\nGlobal prefix: /scim (configurable via API_PREFIX). URL rewriting: /scim/v2/* → /scim/*.",
    "contact": { "name": "SCIMServer", "url": "https://github.com/pranems/SCIMServer" },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000/scim", "description": "Local development (default)" },
    { "url": "http://localhost:6000/scim", "description": "Local (port 6000)" },
    { "url": "http://localhost:8080/scim", "description": "Docker" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Shared secret, per-endpoint credential, or OAuth JWT" },
      "oauthJwt": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "OAuth 2.0 JWT with sub claim - required for /Me" }
    },
    "schemas": {
...
      "LogConfig": { "type": "object", "properties": { "globalLevel": { "type": "string", "enum": ["TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL", "OFF"] }, "categoryLevels": { "type": "object" }, "endpointLevels": { "type": "object" }, "includePayloads": { "type": "boolean" }, "includeStackTraces": { "type": "boolean" }, "maxPayloadSizeBytes": { "type": "integer" }, "format": { "type": "string" }, "availableLevels": { "type": "array", "items": { "type": "string" } }, "availableCategories": { "type": "array", "items": { "type": "string" } } } },
...
    },
    "parameters": {
      "endpointId": { "name": "endpointId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Endpoint (tenant) UUID" },
...
    }
  },
  "paths": {
...
    "/admin/logs": {
      "get": { "tags": ["Admin - General"], "summary": "List request logs", "operationId": "listLogs", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "page", "in": "query", "schema": { "type": "integer" } }, { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }, { "name": "method", "in": "query", "schema": { "type": "string" } }, { "name": "status", "in": "query", "schema": { "type": "string" } }, { "name": "hasError", "in": "query", "schema": { "type": "boolean" } }, { "name": "urlContains", "in": "query", "schema": { "type": "string" } }, { "name": "since", "in": "query", "schema": { "type": "string", "format": "date-time" } }, { "name": "until", "in": "query", "schema": { "type": "string", "format": "date-time" } }, { "name": "search", "in": "query", "schema": { "type": "string" } }, { "name": "includeAdmin", "in": "query", "schema": { "type": "boolean" } }, { "name": "hideKeepalive", "in": "query", "schema": { "type": "boolean" } }], "responses": { "200": { "description": "Paginated logs" } } }
    },
...
    "/admin/log-config": {
      "get": { "tags": ["Admin - Log Config"], "summary": "Get log configuration", "operationId": "getLogConfig", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Config", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LogConfig" } } } } } },
      "put": { "tags": ["Admin - Log Config"], "summary": "Update log configuration", "operationId": "updateLogConfig", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "globalLevel": { "type": "string" }, "includePayloads": { "type": "boolean" }, "includeStackTraces": { "type": "boolean" }, "maxPayloadSizeBytes": { "type": "integer" }, "format": { "type": "string" }, "categoryLevels": { "type": "object" } } } } } }, "responses": { "200": { "description": "Updated" } } }
    },
    "/admin/log-config/level/{level}": {
      "put": { "tags": ["Admin - Log Config"], "summary": "Quick-set global log level", "operationId": "setGlobalLogLevel", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "level", "in": "path", "required": true, "schema": { "type": "string", "enum": ["TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL", "OFF"] } }], "responses": { "200": { "description": "Level set" } } }
    },
    "/admin/log-config/category/{category}/{level}": {
      "put": { "tags": ["Admin - Log Config"], "summary": "Set category log level", "operationId": "setCategoryLogLevel", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "category", "in": "path", "required": true, "schema": { "type": "string", "enum": ["http", "auth", "scim.user", "scim.group", "scim.patch", "scim.filter", "scim.discovery", "endpoint", "database", "oauth", "general"] } }, { "name": "level", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Set" } } }
    },
    "/admin/log-config/endpoint/{endpointId}/{level}": {
      "put": { "tags": ["Admin - Log Config"], "summary": "Set endpoint log level", "operationId": "setEndpointLogLevel", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "level", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Set" } } }
    },
    "/admin/log-config/endpoint/{endpointId}": {
      "delete": { "tags": ["Admin - Log Config"], "summary": "Clear endpoint log level override", "operationId": "clearEndpointLogLevel", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }], "responses": { "204": { "description": "Cleared" } } }
    },
    "/admin/log-config/recent": {
      "get": { "tags": ["Admin - Log Config"], "summary": "Get recent in-memory logs", "operationId": "getRecentLogs", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer" } }, { "name": "level", "in": "query", "schema": { "type": "string" } }, { "name": "category", "in": "query", "schema": { "type": "string" } }, { "name": "requestId", "in": "query", "schema": { "type": "string" } }, { "name": "endpointId", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Recent logs" } } },
      "delete": { "tags": ["Admin - Log Config"], "summary": "Clear log buffer", "operationId": "clearRecentLogs", "security": [{ "bearerAuth": [] }], "responses": { "204": { "description": "Cleared" } } }
    },
    "/admin/log-config/stream": {
      "get": { "tags": ["Admin - Log Config"], "summary": "Real-time log stream (SSE)", "operationId": "streamLogs", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "level", "in": "query", "schema": { "type": "string" } }, { "name": "category", "in": "query", "schema": { "type": "string" } }, { "name": "endpointId", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "SSE stream", "content": { "text/event-stream": { "schema": { "type": "string" } } } } } }
    },
    "/admin/log-config/download": {
      "get": { "tags": ["Admin - Log Config"], "summary": "Download logs as file", "operationId": "downloadLogs", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "format", "in": "query", "schema": { "type": "string", "enum": ["json", "ndjson"] } }, { "name": "limit", "in": "query", "schema": { "type": "integer" } }, { "name": "level", "in": "query", "schema": { "type": "string" } }, { "name": "category", "in": "query", "schema": { "type": "string" } }, { "name": "requestId", "in": "query", "schema": { "type": "string" } }, { "name": "endpointId", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Log file" } } }
    },
...
  }
}